requirements.txt
# 核心依赖
langchain>=0.1.0
langchain-core>=0.3.0
langchain-openai>=0.1.0
langchain-community>=0.1.0

# 大语言模型
openai>=1.0.0
zhipuai>=2.0.0

# 向量化和嵌入
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# 文档处理
unstructured>=0.10.0
markdown>=3.4.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# 数据采集
pandas>=2.0.0
numpy>=1.24.0

# 部署框架
gradio>=4.0.0
streamlit>=1.28.0

# 工具库
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0

# 文本处理
jieba>=0.42.1
transformers>=4.30.0
torch>=2.0.0

# 其他工具
lxml>=4.9.0
html5lib>=1.1 